+++ go/src/github.com/bitrise-core/bitrise-init/_tests/integration/reactnative_test.go
@@ -93,9 +93,29 @@ var sampleAppsReactNativeSubdirResultYML = fmt.Sprintf(`options:
                 env_key: BITRISE_SCHEME
                 value_map:
                   SampleAppsReactNativeAndroid:
-                    config: react-native-android-ios-test-config
+                    title: ipa export method
+                    env_key: BITRISE_EXPORT_METHOD
+                    value_map:
+                      ad-hoc:
+                        config: react-native-android-ios-test-config
+                      app-store:
+                        config: react-native-android-ios-test-config
+                      development:
+                        config: react-native-android-ios-test-config
+                      enterprise:
+                        config: react-native-android-ios-test-config
                   SampleAppsReactNativeAndroid-tvOS:
-                    config: react-native-android-ios-test-config
+                    title: ipa export method
+                    env_key: BITRISE_EXPORT_METHOD
+                    value_map:
+                      ad-hoc:
+                        config: react-native-android-ios-test-config
+                      app-store:
+                        config: react-native-android-ios-test-config
+                      development:
+                        config: react-native-android-ios-test-config
+                      enterprise:
+                        config: react-native-android-ios-test-config
 configs:
   react-native:
     react-native-android-ios-test-config: |
@@ -130,6 +150,7 @@ configs:
               inputs:
               - project_path: $BITRISE_PROJECT_PATH
               - scheme: $BITRISE_SCHEME
+              - export_method: $BITRISE_EXPORT_METHOD
               - configuration: Release
           - deploy-to-bitrise-io@%s: {}
         primary:
@@ -191,9 +212,29 @@ var sampleAppsReactNativeIosAndAndroidResultYML = fmt.Sprintf(`options:
                 env_key: BITRISE_SCHEME
                 value_map:
                   SampleAppsReactNativeAndroid:
-                    config: react-native-android-ios-test-config
+                    title: ipa export method
+                    env_key: BITRISE_EXPORT_METHOD
+                    value_map:
+                      ad-hoc:
+                        config: react-native-android-ios-test-config
+                      app-store:
+                        config: react-native-android-ios-test-config
+                      development:
+                        config: react-native-android-ios-test-config
+                      enterprise:
+                        config: react-native-android-ios-test-config
                   SampleAppsReactNativeAndroid-tvOS:
-                    config: react-native-android-ios-test-config
+                    title: ipa export method
+                    env_key: BITRISE_EXPORT_METHOD
+                    value_map:
+                      ad-hoc:
+                        config: react-native-android-ios-test-config
+                      app-store:
+                        config: react-native-android-ios-test-config
+                      development:
+                        config: react-native-android-ios-test-config
+                      enterprise:
+                        config: react-native-android-ios-test-config
 configs:
   react-native:
     react-native-android-ios-test-config: |
@@ -227,6 +268,7 @@ configs:
               inputs:
               - project_path: $BITRISE_PROJECT_PATH
               - scheme: $BITRISE_SCHEME
+              - export_method: $BITRISE_EXPORT_METHOD
               - configuration: Release
           - deploy-to-bitrise-io@%s: {}
         primary: